AddCity classes moved to addCityFragment package
