+++ app/src/main/java/com/simplertutorials/android/wheathograophy/ui/MainActivity.kt
@@ -6,7 +6,7 @@ import androidx.fragment.app.Fragment
 import com.simplertutorials.android.wheathograophy.R
 import com.simplertutorials.android.wheathograophy.ui.fragments.CityListFragment
 import com.simplertutorials.android.wheathograophy.ui.fragments.WeatherInfoFragment
-import com.simplertutorials.android.wheathograophy.ui.fragments.AddCityFragment
+import com.simplertutorials.android.wheathograophy.ui.fragments.addCityFragment.AddCityFragment
 
 class MainActivity : AppCompatActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
