+++ app/src/main/java/com/simplertutorials/android/wheathograophy/ui/fragments/addCityFragment/AddCityFragmentPresenter.kt
@@ -1,10 +1,11 @@
-package com.simplertutorials.android.wheathograophy.ui.fragments
+package com.simplertutorials.android.wheathograophy.ui.fragments.addCityFragment
 
 import com.simplertutorials.android.wheathograophy.data.database.DatabaseRepository
 import com.simplertutorials.android.wheathograophy.domain.City
 
 class AddCityFragmentPresenter(private val databaseRepository: DatabaseRepository,
-                               private val view: AddCityFragment) {
+                               private val view: AddCityFragment
+) {
 
     fun saveCity(cityName: String) {
         //save the city, show the info snackBar and return to the CityListFragment
